Template for facilities.json and rolled discriminators into facility
