+++ server/models/gov/facility/facility.js
@@ -29,4 +29,29 @@ function validateFacility(facility) {
   return Joi.validate(facility, schema, { "allowUnknown": true });
 }
 
-module.exports = { Facility, validateFacility }
+const Lab = Facility.discriminator('Lab', new Schema({
+  type: { type: String, min: 2, maxlength: 50, default: 'Lab'},
+  sciRate: { type: Number, default: 0 },
+  bonus: { type: Number, default: 0 },
+  funding: { type: Number, default: 0 },
+  research: [{ type: Schema.Types.ObjectId, ref: 'Research' }]
+}));
+
+const Hanger = Facility.discriminator('Hanger', new Schema({
+  type: { type: String, min: 2, maxlength: 50, default: 'Hanger'}
+}));
+
+const Factory = Facility.discriminator('Factory', new Schema({
+  type: { type: String, min: 2, maxlength: 50, default: 'Factory'},
+  project: [{ type: Schema.Types.ObjectId, ref: 'Equipment' }]
+}));
+
+const Crisis = Facility.discriminator('Crisis', new Schema({
+  type: { type: String, min: 2, maxlength: 50, default: 'Crisis'}
+}));
+
+const Civilian = Facility.discriminator('Civilian', new Schema({
+  type: { type: String, min: 2, maxlength: 50, default: 'Civilian'}
+}));
+
+module.exports = { Facility, validateFacility, Lab, Hanger, Factory, Crisis, Civilian }
